requirements.txt
async-generator==1.10
attrs==23.1.0
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
exceptiongroup==1.1.1
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==6.0.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
MouseInfo==0.1.3
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
phonenumbers==8.13.19
pillow==10.2.0
plyer==2.1.0
protobuf==4.25.3
psutil==5.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycountry==22.3.5
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
StrEnum==0.4.15
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
webdriver-manager==4.0.0
wsproto==1.2.0
